`fn mask_rust`: Restore support for negative strides

+++ src/mc.rs
@@ -710,7 +710,7 @@ unsafe fn w_avg_rust<BD: BitDepth>(
 
 unsafe fn mask_rust<BD: BitDepth>(
     mut dst: *mut BD::Pixel,
-    dst_stride: usize,
+    dst_stride: isize,
     tmp1: &[i16; COMPINTER_LEN],
     tmp2: &[i16; COMPINTER_LEN],
     w: usize,
@@ -737,7 +737,7 @@ unsafe fn mask_rust<BD: BitDepth>(
         tmp1 = &tmp1[w..];
         tmp2 = &tmp2[w..];
         mask = mask.offset(w as isize);
-        dst = dst.offset(dst_stride as isize);
+        dst = dst.offset(dst_stride);
     }
 }
 
@@ -1722,7 +1722,7 @@ unsafe extern "C" fn mask_c_erased<BD: BitDepth>(
 ) {
     mask_rust(
         dst.cast(),
-        dst_stride as usize,
+        dst_stride,
         tmp1,
         tmp2,
         w as usize,
